requirements.txt
django==5.1.6
djangorestframework==3.15.2
django-environ==0.12.0
requests==2.32.3
beautifulsoup4==4.13.3